auth/Crypto: fix TAINTED_SCALAR
Make sure we pass only size_t to get_random_bytes() as size parameter
instead of int to prevent usage of a tainted variable when e.g. a
negative value is passed to read() for nbyte.

Fix for:
CID 1219616 (#1 of 1): Use of untrusted scalar value (TAINTED_SCALAR)
 tainted_data: Passing tainted variable buf to a tainted sink.
 tainted_data_argument: Calling function cls_gen_random_bytes taints argument buf.
CID 1219617 (#1 of 1): Use of untrusted scalar value (TAINTED_SCALAR)
 tainted_data_argument: Calling function get_random_bytes taints argument buf.
 tainted_data: Passing tainted variable buf to a tainted sink.

Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>